requirements.txt
asgiref==3.11.0
dj-database-url==2.1.0
Django==6.0.1
gunicorn==23.0.0
pillow==12.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.5
stripe==10.4.0
whitenoise==6.6.0
